[accessibility] Extension toolbar and infobar container views should not have state visible if they are not rendered on screen.
Add the ROLE_PANE role to viewsaccessibility types.

Provide accessible role and name for BrowserActionsContainer, InfoBarContainer, and InfoBar.

Set BrowserActionsContainer view visibility to true only when it contains browser actions.

Set ExtensionShelf vew visibility to true only when it has non-zero height.

BUG=36224
TEST=Verify extension toolbar and infobar container accessibility info within AccExplore32.exe

Review URL: http://codereview.chromium.org/669275

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43044 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8a0a55e61c5efd4711fbf21d49b4001a4cc9908b

+++ views/accessibility/view_accessibility.h
@@ -130,11 +130,11 @@ class ATL_NO_VTABLE ViewAccessibility
 
   // Returns a conversion from the Role (as defined in accessibility_types.h)
   // to an MSAA role.
-  long MSAARole(AccessibilityTypes::Role role);
+  int32 MSAARole(AccessibilityTypes::Role role);
 
   // Returns a conversion from the State (as defined in accessibility_types.h)
   // to MSAA states set.
-  long MSAAState(AccessibilityTypes::State state);
+  int32 MSAAState(AccessibilityTypes::State state);
 
   // Returns the IAccessible interface for a native view if applicable.
   // Returns S_OK on success.
